refactor: split oversized protocol and catalog helpers
