+++ packages/gateway-protocol/src/index.ts
@@ -7,10 +7,10 @@ export {
   type ClawHubTrustErrorCode,
   type ClawHubTrustErrorDetails,
 } from "./clawhub-trust-error-details.js";
-import type { Static, TSchema } from "typebox";
-import { Compile, type Validator as TypeBoxValidator } from "typebox/compile";
 import type { ValidationError } from "./validation-errors.js";
 export { formatValidationErrors, type ValidationError } from "./validation-errors.js";
+import { lazyCompile } from "./protocol-validator.js";
+export type { ProtocolValidator } from "./protocol-validator.js";
 export * from "./schema/worker-inference.js";
 export type {
   SessionCatalog,
@@ -487,67 +487,6 @@ import {
   FsListDirParamsSchema,
   FsListDirResultSchema,
 } from "./schema.js";
-
-/** Runtime validator shape shared by gateway clients and server handlers. */
-export type ProtocolValidator<T = unknown> = ((data: unknown) => data is T) & {
-  errors: ValidationError[] | null; // Ajv-style last validation errors.
-  /** Original schema used by the validator, exposed for diagnostics/tests. */
-  schema: unknown;
-};
-
-// Defer TypeBox compilation because this module is common on startup paths.
-function lazyCompile<const Schema extends TSchema>(
-  schema: Schema,
-  precheck?: (data: unknown) => ValidationError | undefined,
-): ProtocolValidator<Static<Schema>>;
-// Keep compact hand-authored public types where schema-derived declarations are intentionally avoided.
-function lazyCompile<T>(
-  schema: TSchema,
-  precheck?: (data: unknown) => ValidationError | undefined,
-): ProtocolValidator<T>;
-function lazyCompile<T = unknown>(
-  schema: TSchema,
-  precheck?: (data: unknown) => ValidationError | undefined,
-): ProtocolValidator<T> {
-  let compiled: TypeBoxValidator | undefined;
-  let errors: ValidationError[] | null = null;
-
-  const getCompiled = () => {
-    compiled ??= Compile(schema as never);
-    return compiled;
-  };
-
-  const validate = ((data: unknown): data is T => {
-    const precheckError = precheck?.(data);
-    if (precheckError) {
-      errors = [precheckError];
-      return false;
-    }
-    const current = getCompiled();
-    const valid = current.Check(data);
-    errors = valid ? null : ([...current.Errors(data)] as ValidationError[]);
-    return valid;
-  }) as ProtocolValidator<T>;
-
-  Object.defineProperties(validate, {
-    errors: {
-      configurable: true,
-      enumerable: true,
-      get: () => errors,
-      set: (nextErrors: ValidationError[] | null | undefined) => {
-        // Preserve Ajv-compatible mutability for callers/tests that clear errors.
-        errors = nextErrors ?? null;
-      },
-    },
-    schema: {
-      configurable: true,
-      enumerable: true,
-      get: () => schema,
-    },
-  });
-
-  return validate;
-}
 
 // Validator names mirror schemas so callers can pair them with wire contracts.
 export const validateCommandsListParams = lazyCompile(CommandsListParamsSchema);
